### 2020-10-16 #2755 @danielskatz
@kanishkan91 - thanks for your submission.

In my opinion, the paper is too long.  I suggest moving some of the example content into the software documentation, perhaps in a tutorial on usage.

I also suggest that most of the information just before the acknowledgements could either be a few lines of text, rather than a page of bullets, or again, moved into the repository/documentation/README.

Finally, note that whedon has found what it thinks might be some missing DOIs. If these are correct, please add them to your bib file.

Use ```@whedon generate pdf``` and ```@whedon check references``` after you make changes to check the results - each command needs to be the first (or only) thing in a new comment.
