### 2020-10-16 #2755 @danielskatz
Also, you probably need to protect some letters in the bib with {} to make/keep them upper case, such as FAIR in one reference.
