### 2020-10-16 #2755 @kanishkan91
@danielskatz Thanks much for the comments! we will work on cutting down the length of the paper as you suggest.

By the way, these examples are documented on a user tutorial here on the repository - https://jgcri.github.io/ambrosia/articles/ambrosia_vignette.html

This is also referenced in the README.md . The user tutorial in addition to other documentation can also be accessed here-
https://jgcri.github.io/ambrosia/

Do let us know what you think. Thanks again!
