### 2020-10-19 ce36345
Updates to paper
