### 2020-10-22 fd6452e
Updates to documentation, paper
