+++ paper/paper.md
@@ -97,7 +97,7 @@ where, $y_{i}$ is the $i$th data value and $\hat{y_{i}}$ is the corresponding mo
 
 By applying the 11 parameters to the equations described above, the user can generate estimates of demand for staples and non-staple commodities in thousand calories  across different income levels and prices.     
 
-# Functions and customization
+# Main functions and customization
 
 The [```ambrosia_vignette.rmd```](https://jgcri.github.io/ambrosia/articles/ambrosia_vignette.html) provides usable examples for the major functions within the code. As described in the summary statement, the functions within `ambrosia` can be classified into three distinct categories:
 
@@ -105,174 +105,17 @@ The [```ambrosia_vignette.rmd```](https://jgcri.github.io/ambrosia/articles/ambr
 2) functions to estimate parameters using custom data,
 3) functions to process raw data for parameter estimation.
 
-The code contains a number of helper functions that may be useful to the user depending upon their research requirements.
 In addition to this, an interactive version of the food demand model can be launched by the user through the ```runapp()``` function to explore the impact of different parameters.
 
-## Functions to explore demand variables
+The `ambrosia` package can be easily loaded as a standard R package after installation from GitHub. The user can calculate demand for staples and non-staples using the  ```food.dmnd()``` function. The user will have to pass in a dataset with the price of staples ($Ps$), price of non-staples ($Pn$), incomes ($Y$) in GDP per capita in thousand USD. In addition to the dataset, the user must pass a vector of 11 parameters. In order to functionalize the parameters, the code contains a function called ```vec2param()``` that will generate a parameter structure that can be used by the food demand function. The usage of the functions are described in the example below. The example makes use of the vector of parameters directly from @edmonds2017global. The food demand function is implemented using equations (1), (2), (3) described above. The user can also calculate and analyze price elasticities. These elasticities are calculated in accordance with equation (4) described above. The function ```calc1eps()``` can be used to recalculate price elasticities
 
-The `ambrosia` package can be easily loaded as a standard R package after installation from GitHub. The user can calculate demand for staples and non-staples using the  ```food.dmnd()``` function. The user will have to pass in a dataset with the price of staples ($Ps$), price of non-staples ($Pn$), incomes ($Y$) in GDP per capita in thousand USD. In addition to the dataset, the user must pass a vector of 11 parameters. In order to functionalize the parameters, the code contains a function called ```vec2param()``` that will generate a parameter structure that can be used by the food demand function. The usage of the functions are described in the example below. The example makes use of the vector of parameters directly from @edmonds2017global. The food demand function is implemented using equations (1), (2), (3) described above.
-
-### Example 1: Creating estimates of demand
-```R
-# Parameter values are taken from Edmonds et al. (Pg 12, Table 3)
-param_vector <- c(1.28, 1.14, -0.19, 0.21, -0.33, 0.5, 0.1, 16, 5.06, 100, 20)
-
-# Names of the parameters above are as follows.
-# These are the same as the names in Table 1 in Appendix I  
-parameter_names <- c('A_s', 'A_n', 'xi_ss', 'xi_cross', 'xi_nn', 'nu1_n',
-                     'lambda_s', 'k_s', 'Pm', 'psscl','pnscl')
-
-# Generate parameter set for the food demand model
-params <- vec2param(param_vector)                     
-
-# Create a sample dataset
-Test_Data <- data.frame(Y=seq(0.1, 30, by=0.1))
-Test_Data %>% mutate(Ps=0.1, Pn=0.2) -> Test_Data
-
-# Calculate food demand
-Food_Demand<-food.dmnd(Ps = Test_Data$Ps,
-                       Pn = Test_Data$Pn,
-                       Y = Test_Data$Y,
-                       params = params)
-```
-
[…]
