+++ R/calculate_parameters.R
@@ -3,6 +3,18 @@
 
 #' Calculate the 11 parameters for ambrosia using data calculated by process_food_demand_data() by maximizing log likelihood
 #'
+#'@details The following steps are involved in the parameter estimation function,
+#'
+#'1) First a log-likelihood function is set up with the data returned by the function above.
+#'
+#'2) Next, the value returned by this function will be maximized using ```optim()```. The user can provide
+#'  a seed of initial parameters to begin the optimization process (the lowest possible seed would be the lowest
+#'  values of all 11 parameters). The default seed is set to the original parameters from Edmonds et al. The user
+#'  can now specify the optimization method to be used. The default is set to the "BFGS" method, but the user can
+#'  also run the optimization using methods such as "Neldor-Mead".
+#'
+#'3) Finally, the function will now return a vector of parameters that can be used to derive estimates of food demand (similar to Example 1 above). The function also prints out the maximized value of the log-likelihood function, so that the user can verify the efficiency and effectiveness of the parameter estimation.
+#'
 #' @return A vector with the 11 parameters which are as follows,
 #'
 #' A_s : A scaling term for staple commodities
@@ -85,7 +97,29 @@ calculate_ambrosia_params <- function(optim_method = "BFGS",
 
 }
 
-#' Create dataset with observational error for log likelihood calculation using clustering.
+#'Create dataset with observational error for log likelihood calculation using clustering.
+#'
+#'
+#'There are a few steps that the function will perform on a sample dataset (see details).
+#'
+#' The steps that will be performed by the function are :
+#'
+#'1) It will ensure that the user's dataset contains all columns required for parameter estimation.
+#'
+#'2) It will filter out anomalies and outliers using parameterized cutoff values selected by the user.
+#'   This step is necessary since data on food consumption and prices are often incomplete which may lead to
+#'   unrealistically high or low values of consumption or prices in the dataset.
+#'
+#'3) After this, the function will create clusters of observations from the dataset based on income levels,
+#'   and prices of staples and non-staples. This step is necessary because this being economic data,
+#'   the observational error can only be calculated within different clusters. The code will also check for a  user
+#'   specified minimum number of clusters (if there are anomalies within the dataset, the clustering can be incorrect leading in a small number of clusters).
+#'   The clustering is implemented using the Divisive Analysis Clustering Algorithm (DIANA).
+#'
+#'4) Once the clustering is completed, the code will calculate the observational error
+#'   which is the variance in food demand for staples and non-staples .Note that the user can chose a
+#'   lower limit on the observational error calculated. The default value of the lower limit is 0.01.
+#'
 #'
 #' @return A dataframe  called Processed_Data_for_MC with the following columns
 #'
@@ -101,15 +135,15 @@ calculate_ambrosia_params <- function(optim_method = "BFGS",
 #' @param optim_method The optimization method to be used for maximization of the log likelihood. The default is set to BFGS
 #' @param data A data.frame or data.table with the raw data. Data should contain following names,
 #'
-#' s_cal_pcap_day_thous (Containing 1000 calories per capita per day for staples)
+#' {s_cal_pcap_day_thous} (Containing 1000 calories per capita per day for staples)
 #'
-#' ns_cal_pcap_day_thous (Containing 1000 calories per capita per day for non-staples)
+#' {ns_cal_pcap_day_thous} (Containing 1000 calories per capita per day for non-staples)
 #'
-#' gdp_pcap_thous (Containing GDP per capita in thous USD)
+#' {gdp_pcap_thous} (Containing GDP per capita)
 #'
-#' s_usd_p1000cal (Price of 1000 calories for staples per person per day)
[…]
