+++ man/calculate_ambrosia_params.Rd
@@ -63,6 +63,18 @@ pnscl : Price scaling parameter for staples that is applied to the Price of non-
 \description{
 Calculate the 11 parameters for ambrosia using data calculated by process_food_demand_data() by maximizing log likelihood
 }
+\details{
+The following steps are involved in the parameter estimation function,
+1) First a log-likelihood function is set up with the data returned by the function above.
+
+2) Next, the value returned by this function will be maximized using ```optim()```. The user can provide
+ a seed of initial parameters to begin the optimization process (the lowest possible seed would be the lowest
+ values of all 11 parameters). The default seed is set to the original parameters from Edmonds et al. The user
+ can now specify the optimization method to be used. The default is set to the "BFGS" method, but the user can
+ also run the optimization using methods such as "Neldor-Mead".
+
+3) Finally, the function will now return a vector of parameters that can be used to derive estimates of food demand (similar to Example 1 above). The function also prints out the maximized value of the log-likelihood function, so that the user can verify the efficiency and effectiveness of the parameter estimation.
+}
 \author{
 KBN 2020
 }
