### 2020-10-22 6cf72f9
Update paper.md
