+++ paper/paper.md
@@ -40,9 +40,7 @@ affiliations:
 The `ambrosia` R package was developed to calculate food demand for staples and non-staple commodities that is responsive to changing levels of incomes and prices. `ambrosia` implements the framework to quantify food demand as established by @edmonds2017global and allows the user to explore and estimate different variables related to the food demand system. Currently `ambrosia` provides three main functions:
 
 (1)	calculation of food demand for any given set of parameters including income levels and prices;
-
 (2)	estimation of parameters within a given a dataset.  Note:  `ambrosia` is used to calculate parameters for the food demand model implemented in the Global Change Analysis Model [GCAM; @calvin2019gcam];
-
 (3)	exploration and preparation of raw data before starting a parameter estimation.
 
 # Statement of need
