+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={04},
   pages={1750012},
   year={2017},
-  publisher={World Scientific}
+  publisher={World Scientific},
+  DOI={10.1142/s2010007817500129}
 }
 @wikidocumentation{calvinetal2016global,
   title={GCAM v5.3 Documentation: Global Change Analysis Model (GCAM)},
@@ -32,7 +33,8 @@
   volume={8},
   number={PNNL-SA-142149},
   year={2020},
-  publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)}
+  publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)},
+  DOI={10.5334/jors.266}
 }
 @article{hartin2015hector,
   title={A simple object-oriented and open-source model for scientific and policy analyses of the global climate system – Hector v1.0},
@@ -41,7 +43,8 @@
   volume={8},
   number={gmd-8-939-2015},
   year={2015},
-  publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)}
+  publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)},
+  DOI = {10.5194/gmd-8-939-2015}
 }
 @article{li2017xanthos,
   title={Xanthos--a global hydrologic model},
@@ -50,7 +53,8 @@
   volume={5},
   number={PNNL-SA-126584},
   year={2017},
-  publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)}
+  publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)},
+  DOI={10.5334/jors.181}
 }
 @article{vernon2018demeter,
   title={Demeter--A Land Use and Land Cover Change Disaggregation Model},
@@ -67,7 +71,8 @@
   journal={CRC Press},
   volume={30},
   year={1988},
-  publisher={CRC Press}
+  publisher={CRC Press},
+  DOI={10.1201/9780203735268}
 }
 @article{hastings1970,
   title={ Monte Carlo sampling methods using Markov chains and their applications},
@@ -75,10 +80,11 @@
   journal={Biometrika},
   volume={1},
   year={1988},
-  publisher={Biometrika}
+  publisher={Biometrika},
+  DOI={10.1093/biomet/57.1.97}
 }
 @article{wilkinson2016fair,
-  title={The FAIR Guiding Principles for scientific data management and stewardship},
+  title={The {FAIR} Guiding Principles for scientific data management and stewardship},
   author={Wilkinson, Mark D and Dumontier, Michel and Aalbersberg, IJsbrand Jan and Appleton, Gabrielle and Axton, Myles and Baak, Arie and Blomberg, Niklas and Boiten, Jan-Willem and da Silva Santos, Luiz Bonino and Bourne, Philip E and others},
   journal={Scientific data},
   volume={3},
@@ -95,14 +101,16 @@
   number={1},
   pages={94-99},
   year={1969},
-  publisher={The Computer Journal}
+  publisher={The Computer Journal},
+  DOI={10.1093/comjnl/12.1.94}
 }
 @article{kaufman2009,
   title={Finding groups in data: an introduction to cluster analysis},
   author={Kaufman, Leonard, and Peter J. Rousseeuw},
   volume={344},
   year={2009},
-  publisher={John Wiley & Sons}
+  publisher={John Wiley & Sons},
+  DOI={10.2307/1269576}
 }
 @misc{pnnl_2020,
     url={https://gcims.pnnl.gov/},
@@ -118,7 +126,8 @@
   volume={4},
   number={7},
   pages={2928},
-  year={2009}
+  year={2009},
[…]
