### 2020-10-22 db15d45
Updates to function documentation, paper
