+++ R/food-demand.R
@@ -62,13 +62,13 @@
 #' @export
 food.dmnd <- function(Ps, Pn, Y, params, rgn=NULL)
 {
-    Pm <- params$Pm
-    psscl <- params$psscl
-    pnscl <- params$pnscl
-    ## Normalize income and prices to Pm
-    Ps <- Ps/Pm * psscl
-    Pn <- Pn/Pm * pnscl
-    Y  <- Y/Pm
+  Pm <- params$Pm
+  psscl <- params$psscl
+  pnscl <- params$pnscl
+  ## Normalize income and prices to Pm
+  Ps <- Ps/Pm * psscl
+  Pn <- Pn/Pm * pnscl
+  Y  <- Y/Pm
 
   # get eta values
   eta.s <- params$yfunc[[1]](Y,FALSE)
@@ -110,14 +110,14 @@ food.dmnd <- function(Ps, Pn, Y, params, rgn=NULL)
   qs <- qvals[1,]
   qn <- qvals[2,]
   ## calculate Qm as the budget residual.
-    resid <- Y - (Ps*qs/psscl + Pn*qn/pnscl)
-    qm <-  resid / Pm
-    alpha.m <- resid / Y
+  resid <- Y - (Ps*qs/psscl + Pn*qn/pnscl)
+  qm <-  resid / Pm
+  alpha.m <- resid / Y
 
-    if(is.null(rgn))
-        data.frame(Qs=qs, Qn=qn, Qm=qm, alpha.s=alpharslt[1,], alpha.n=alpharslt[2,], alpha.m=alpha.m)
-    else
-        data.frame(Qs=qs, Qn=qn, Qm=qm, alpha.s=alpharslt[1,], alpha.n=alpharslt[2,], alpha.m=alpha.m, rgn=rgn)
+  if(is.null(rgn))
+    data.frame(Qs=qs, Qn=qn, Qm=qm, alpha.s=alpharslt[1,], alpha.n=alpharslt[2,], alpha.m=alpha.m)
+  else
+    data.frame(Qs=qs, Qn=qn, Qm=qm, alpha.s=alpharslt[1,], alpha.n=alpharslt[2,], alpha.m=alpha.m, rgn=rgn)
 }
 
 #' Calculate the exponents in the demand equation.
@@ -135,17 +135,17 @@ food.dmnd <- function(Ps, Pn, Y, params, rgn=NULL)
 #' @export
 calc1eps <- function(alpha.s, alpha.n, eta.s, eta.n, xi) {
 
-    ## First apply symmetry condition.  This means that the xi.sn
-    ## value will be ignored.  Also, set a floor on the terms to
-    ## ensure that the function is well-behaved.
-    alphamin <- 0.1
-    xi[3] <- max(alpha.n, alphamin)/max(alpha.s,alphamin) * xi[2]
+  ## First apply symmetry condition.  This means that the xi.sn
+  ## value will be ignored.  Also, set a floor on the terms to
+  ## ensure that the function is well-behaved.
+  alphamin <- 0.1
+  xi[3] <- max(alpha.n, alphamin)/max(alpha.s,alphamin) * xi[2]
 
-    ## Now calculate the epsilon matrix using the Slutsky equation.
-    c(xi[1] - alpha.s * eta.s,    # ess
-      xi[2] - alpha.s * eta.n,    # ens
-      xi[3] - alpha.n * eta.s,    # esn
-      xi[4] - alpha.n * eta.n)   # enn
+  ## Now calculate the epsilon matrix using the Slutsky equation.
+  c(xi[1] - alpha.s * eta.s,    # ess
[…]
