+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   ambrosia_vignette: ambrosia_vignette.html
-last_built: 2020-10-22T03:58Z
+last_built: 2020-10-22T17:55Z
 
