### 2020-10-23 3ed9e31
Updates to paper
