+++ man/compute_bias_corrections.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/food-demand.R
-\name{compute.bias.corrections}
-\alias{compute.bias.corrections}
+\name{compute_bias_corrections}
+\alias{compute_bias_corrections}
 \title{Compute regional bias corrections.}
 \usage{
-compute.bias.corrections(params, obs.trn)
+compute_bias_corrections(params, obs.trn)
 }
 \arguments{
 \item{params}{Model parameter structure (described in
