+++ NAMESPACE
@@ -11,7 +11,7 @@ export(calc.pop.weight)
 export(calc1eps)
 export(calculate_ambrosia_params)
 export(column.input.table)
-export(compute.bias.corrections)
+export(compute_bias_corrections)
 export(create_dataset_for_parameter_fit)
 export(elasmax)
 export(elasmin)
