### 2020-10-23 #2755 @kanishkan91
Hi all,

As suggested, we have cut down the paper length and shifted some text to the documentation and the user tutorial. We have also fixed and checked the DOI's and addressed other comments. Do let us know if you need anything else from our side. Thanks!
