### 2020-10-28 #2755 @arfon
:wave: @dhhagan - further to @danielskatz's [request from ~10 days ago](https://github.com/openjournals/joss-reviews/issues/2755#issuecomment-710026146) I've gone ahead and assigned you here.
