### 2020-12-06 #2890 @whedon
[REVIEW]: ambrosia: An R package for calculating and analyzing food demand and in accordance with the Edmonds et al. food demand model
**Submitting author:** @kanishkan91 (<a href="http://orcid.org/0000-0001-8483-6216">Kanishka Narayan</a>)
**Repository:** <a href="https://github.com/JGCRI/ambrosia" target ="_blank">https://github.com/JGCRI/ambrosia</a>
**Version:** v1.3.5
