### 2020-12-07 JGCRI/ambrosia#20 @tscheypidi
LICENSE is not BSD-2 / non OSI-approved
The LICENSE file states that the software package is licensed under BSD-2, but the provided license actually extends the original BSD-2 license text (compare the original license text here: https://opensource.org/licenses/BSD-2-Clause) and adds additional burden on the user, making it a custom license.

Please remove the license restrictions which do not belong to the BSD-2 license or change the applied license.

https://github.com/openjournals/joss-reviews/issues/2890
