### 2020-12-10 #2890 @tscheypidi
@kanishkan91 I found two discrepancies between author list of this submission and recorded contributions to the package for which I could not find a direct explanation myself:

- In the commit history Robert Link (@rplzzz) shows up as top contributor of the package (and is 2nd Co-Author of the paper this package relies on) but is not listed as author of this submission. Is that intended and if so, why?
- Ryna Cui is listed as author of this submission but is not showing up in the git history. Can you please explain her contribution to the package?
