### 2020-12-10 #2890 @kanishkan91
@tscheypidi Thanks for the questions. Here are the responses,

- Yes. This was intentional. Robert Link was a co-author on the original Edmonds et al. paper and started the work on the ambrosia R project. But he transitioned to a new role at a different organization early this year and has requested that he no longer be involved with any publications with the current organization given his commitments and requirements at his new organization. Therefore, we want to honor his request.

- Jae Edmonds, Stephanie Waldhoff and Ryna Cui conceptualized the original Edmonds et al. framework and established the scientific validity of the approach (They are all authors on the original science paper). Since their work supports the science of ambrosia, they are included as authors for this publication.

If you would like, we can make the contributions of individual authors explicit in the paper. Let us know what you think. Thanks!
