### 2020-12-14 JGCRI/ambrosia#21 @tscheypidi
install_github with build_vignettes = TRUE?
It seems that by default the supplied vignette is not being installed when running `devtools::install_github('JGCRI/ambrosia')`

I would suggest to ask users to run `devtools::install_github('JGCRI/ambrosia', build_vignettes = TRUE)` instead.

Doing so would also allow to refer to the vignette via the command `vignette("ambrosia_vignette")` instead of referring to the vignettes directory.

(this is related to https://github.com/openjournals/joss-reviews/issues/2890)
