### 2020-12-14 JGCRI/ambrosia#22 @tscheypidi
parameters unclear
As the parameters seem to have a central role in the whole process a bit more explanation and a better visibility of these explanations throughout the documentation would be useful. Some examples

- `A_s` is described as "Scaling parameter for staples (constant)", but it is unclear to me how it works. What exactly is scaled here and which range of scaling factors makes sense? How can these parameters be estimated?
- All parameters lack unit information
- In the vignette the parameter names are introduced and used, but they are not explained. Also a link to a page explaining these parameters is not available.
- vec2param (and probably other documentation pages) are mentioning the parameter names, but not explaining what it is (sidnode: in case of 8 parameters it looks like the documentation is only explaining what the first 7 are, but not what the 8th parameter is).

Please check again the documentation with a focus on the parameters and make sure that the correspoding parameters are either explained on the page or the parameter description is linked along with the use of these parameter names

https://github.com/openjournals/joss-reviews/issues/2890
