### 2020-12-14 JGCRI/ambrosia#22 @kanishkan91
@tscheypidi Thanks much for this comment.  We will work on making the explanations more clear. The problem is that the equations are documented in the paper, the estimation framework is documented in a separate function and parameter bounds are documented separately in another function. We will try to bring this all together meaningfully and get back to you shortly. If you have any other suggestions related to this, do let us know.
