### 2020-12-17 JGCRI/ambrosia#21 @tscheypidi
I would recommend to put the vignette call statement into the "User tutorial and examples" section below. Otherwise, it might be confusing for the user why it is first suggested to call the vignette via `vignette("ambrosia_vignette")` but directly afterwards suggested to navigate to the vignette file using its physical location on the system.
