### 2020-12-17 JGCRI/ambrosia#22 @tscheypidi
based on the vignette my understanding is that the parameters need to be provided in the predefined structure currently created by `vec2param`. Currently, already needs to know which parameters (in which ranges) in which order need to be provided to that function as `x`.

What you could do for better usability but also documentation would be also to allow the provision of all parameters as single arguments. This could be either done in `vec2param` itself as alternative input method, or in another function which also returns a parameter object. It could look something like this:

```
input2param <- function(A_s=1.28, A_n=1.14, xi_ss=-0.19, xi_cross=0.21,
                                     xi_nn=-0.33, nul_n=0.5, lambda_s=0.1,
                                     k_s=16, Pm=5.06,
                                     psscl=100, pnscl=20)
```
or

```
vec2param <- function(x=NULL, A_s=1.28, A_n=1.14, xi_ss=-0.19, xi_cross=0.21,
                                     xi_nn=-0.33, nul_n=0.5, lambda_s=0.1,
                                     k_s=16, Pm=5.06,
                                     psscl=100, pnscl=20)
```
In the latter the single element parameter entries would be ignored if provided via `x`.

Putting each argument explicitly in the function call would have the following advantages:

- The documentation of that function would provide you with a structured setting to properly document each parameter with default values (which could be set as default as shown in the example), meaning, unit and ranges
- You could refer in all other functions to the documentation of this function for explanation of the parameters
- Novice users could create a default parameter object easily via calling `input2param()` or `vec2param()`
