### 2020-12-17 JGCRI/ambrosia#23 @tscheypidi
suggestion for style improvements in vignette
I see some potential for improvement in the provided vignette. In particular, I see the following issues:

- Currently, explanatory text either comes in the form of a header or as a comment in the code snippets. It would improve the readability to convert the lengthy header to short titles of the sections and use some explanatory sentences for each example (in addition to the comments in the code). See for instance `vignette("knitr")` as an example how something like that could look like.
- The flow of the examples is somehow inconsistent in the sense that sometimes an example builds on the calculations of the previous example (e.g. 1.3 continues using the object `Food_Demand` created in 1.2), but in other cases it does not (e.g. 1.1 builds a parameter object, while 1.2 seems to use the same parameter object but rebuilds it from scratch rather than using the result from 1.1). It would be nice to just reuse the objects from previous code snippets.
- As a vignette is typically meant as an exercise that user can easily run the heavily time consuming snippets in 2.1 and 2.4 are in my view not very useful. Wouldn't it be possible for these examples to apply the calculations to dummy data sets and reduce the runtime to a few seconds. Even if this would not provide reasonable results it would teach how the functions need to be called and what objects are returned by it.

Solving these three issues whould in my opinion greatly improve the readability of the vignette.

https://github.com/openjournals/joss-reviews/issues/2890
