### 2020-12-17 JGCRI/ambrosia#24 @tscheypidi
missing reference in README.md
README.md refers to Calvin et al. 2019 but does not provide the corresponding citation.

https://github.com/openjournals/joss-reviews/issues/2890
