### 2020-12-17 JGCRI/ambrosia#25 @tscheypidi
missing contact information in README.md
Guidelines are missing for third parties wishing to 1) Contribute to the software (README states, that one should get in contact with the team, but does not state how to do so) 2) Report issues or problems with the software (should that be done via Github Issues, or differently?)  3) Seek support (unclear who to approach on which channel in this case)

https://github.com/openjournals/joss-reviews/issues/2890
