### 2020-12-17 JGCRI/ambrosia#26 @tscheypidi
shiny app only usable by advanced users
While from a technical point the app works like a charm, it is not really usable for me as a beginner as explanations of the different variables (including their units) are missing.
Without it, its is neither clear what the different parameter settings mean, nor what the tables or figures show.

I would suggest to either add explanations/legends to the app explaining all the different variables and their units (best, but also very time consuming solution) or to put a disclaimer into the documentation of this function and the vignette, explaining that this app is targeted at advanced users and requires a priori knowledge about the meaning of all these variables. Ideally, such a disclaimer should come with an explanation how that knowledge can be obtained.

https://github.com/openjournals/joss-reviews/issues/2890
