### 2021-01-05 19252d2
Updates based on JOSS review
