+++ R/food-demand-plots.R
@@ -75,28 +75,28 @@ make.byyear.plot <- function(byyear.data, pltrgn=NULL)
     }
 
     baseplt <- ggplot2::ggplot(data=modeldata, ggplot2::aes(x=year)) +
-        ggplot2::facet_wrap(~rgn) +
-        ggplot2::ylab('Q (1000 Cal pc/day)') +
-        ggplot2::geom_line(ggplot2::aes(y=value, colour=variable), size=1.5) +
-        ggplot2::geom_ribbon(data=Qs.err, ggplot2::aes(x=year, ymin=Qlo, ymax=Qhi), alpha=0.2) +
-        ggplot2::geom_ribbon(data=Qn.err, ggplot2::aes(x=year, ymin=Qlo, ymax=Qhi), alpha=0.2) +
-        ggplot2::theme(panel.margin=ggplot2::unit(1,'lines'))
+      ggplot2::facet_wrap(~rgn) +
+      ggplot2::ylab('Q (1000 Cal pc/day)') +
+      ggplot2::geom_line(ggplot2::aes(y=value, colour=variable), size=1.5) +
+      ggplot2::geom_ribbon(data=Qs.err, ggplot2::aes(x=year, ymin=Qlo, ymax=Qhi), alpha=0.2) +
+      ggplot2::geom_ribbon(data=Qn.err, ggplot2::aes(x=year, ymin=Qlo, ymax=Qhi), alpha=0.2) +
+      ggplot2::theme(panel.margin=ggplot2::unit(1,'lines'))
 
     ## Add the points for the observed data
     if(is.null(obsdata$obstype)) {
-        ## observation data not separated into training and testing
-        ## sets, so just plot one type of point, but give it a legend
-        ## to make it a little more clear that the dots are
-        ## observations and the lines are model output.
-        obsdata$obstype <- 'obs'
-        baseplt +
-            ggplot2::geom_point(data=obsdata, ggplot2::aes(y=value, colour=variable, shape=obstype)) +
-            ggplot2::scale_shape(guide=ggplot2::guide_legend(title='Observed Data', label=FALSE))
+      ## observation data not separated into training and testing
+      ## sets, so just plot one type of point, but give it a legend
+      ## to make it a little more clear that the dots are
+      ## observations and the lines are model output.
+      obsdata$obstype <- 'obs'
+      baseplt +
+        ggplot2::geom_point(data=obsdata, ggplot2::aes(y=value, colour=variable, shape=obstype)) +
+        ggplot2::scale_shape(guide=ggplot2::guide_legend(title='Observed Data', label=FALSE))
     }
     else {
-        baseplt +
-            ggplot2::geom_point(data=obsdata, ggplot2::aes(y=value, colour=variable, shape=obstype)) +
-            ggplot2::scale_shape(guide=ggplot2::guide_legend(title='Observed Data')) # This one includes labels for the obs types.
+      baseplt +
+        ggplot2::geom_point(data=obsdata, ggplot2::aes(y=value, colour=variable, shape=obstype)) +
+        ggplot2::scale_shape(guide=ggplot2::guide_legend(title='Observed Data')) # This one includes labels for the obs types.
     }
 }
 
@@ -110,14 +110,14 @@ make.byyear.plot <- function(byyear.data, pltrgn=NULL)
 #' @export
 make.byincome.plot <- function(obsdata, params, region=NULL)
 {
-    pcGDP <- value <- region <- variable <- NULL
+  pcGDP <- value <- region <- variable <- NULL
 
-    plotdata <- food.dmnd.byincome(obsdata, params, region)
+  plotdata <- food.dmnd.byincome(obsdata, params, region)
 
[…]
