+++ NAMESPACE
@@ -51,3 +51,8 @@ export(xidefault)
 export(y.vals)
 export(y0.input.box)
 importFrom(dplyr,"%>%")
+importFrom(dplyr,filter)
+importFrom(dplyr,mutate)
+importFrom(dplyr,rename)
+importFrom(stats,optim)
+importFrom(utils,write.csv)

[…]
