### 2021-01-05 f18fdc4
Updated documentation based on JOSS review
