+++ NAMESPACE
@@ -56,3 +56,4 @@ importFrom(dplyr,mutate)
 importFrom(dplyr,rename)
 importFrom(stats,optim)
 importFrom(utils,write.csv)
+importFrom(utils,write.table)
