### 2021-01-05 eb4aff6
Updates to vignette, adding in parameters
