+++ R/calculate_parameters.R
@@ -127,6 +127,7 @@ calculate_ambrosia_params <- function(optim_method = "BFGS",
 #'
 #' sig2Qs- which is the observational error for staples
 #'
+#' @param min_price_pd Minimum price paid for non-staples.
 #' @param min_cal_fd Minimum calories for the food demand model
 #' @param min_clusters Minimum number of clusters to be generated by clustering algoritm. It is recommended to not lowewr this parameter below 20.
 #' @param lower_limit_sigma Lower limit for sigma values calculated
@@ -147,7 +148,7 @@ calculate_ambrosia_params <- function(optim_method = "BFGS",
 #' @importFrom dplyr mutate filter rename
 #' @author KBN 2020
 #' @export
-create_dataset_for_parameter_fit <- function(
+create_dataset_for_parameter_fit <- function(min_price_pd = 20,
                                              min_cal_fd = 1000,
                                              min_clusters = 300,
                                              lower_limit_sigma = 0.01,
