### 2021-01-05 #2890 @kanishkan91
@tscheypidi We have tried to address all of your review comments as a part of release v1.3.1 . I am summarizing our responses to each of the issues below for your convenience,

1.  We have removed the license restrictions as suggested and made sure that this is a BSD2 license.
2.  As suggested, we have recommended explicitly that the user build vignettes when loading the repository. Our instructions also indicate how the user can load the vignette directly.
3.  We have added explicit documentation on the 11 parameters used. These are documented in the documentation of the function `vec2param()` . We have added a table to the documentation that describes each parameter, its units and the acceptable range for the same and how we arrived at the range/how it is recommended that users arrive at the range (95% Joint confidence interval of the range of values generated during the parameter fitting exercise). In addition to this we have made some changes to `vec2param()`   so that the user can either pass in a vector of nameless parameters or change individual parameters or load default parameters (value for each of these parameters is set at separate objects and then combined to a vector)
4. We have incorporated your suggestions for the vignette. You will find that the vignette now contains documentation on what is being done in each section in detail. The code comments are also retained. This way, user can directly use the code or look into more in depth explanations as the case may be.
5. We have taken out the reference in the readme (It made no sense). In its place we have included an explicit hyperlink to GCAM which achieves our desired objective.
6. As suggested, we have added the contact information to the README.md
7. Finally, on the shiny app, the app was intended for new users to experiment with parameter values. As you indicated, in its state it was usable only by experts. We have significantly overhauled the app. We have added documentation for the parameters, the results are also explained and we ended up taking out some technical results that can only be interpreted by experts. We would love to know if the improved app addresses your concerns.

Please do let us know what you think of all of the above. Let us know if we have missed anything important. Thanks once again for this tremendously useful review!
