### 2021-01-05 #2890 @tscheypidi
@kanishkan91 Excellent work and thanks for putting so much effort into addressing my remarks. I particularly appreciate the extensions in vignette and app as well as the extended functionality of `vec2param()`.

@dhhagan I am done with my review and all my concerns have been fully addressed by the authors.
