### 2021-01-13 #2890 @crvernon
Hi @dhhagan !  If we don't hear back from @sahilseth I found some other reviewers from recent JOSS publications that may be able to conduct the review.  I'll not call them out here to ensure they are not tagged until you see fit, hence the quotes around their names:  `@trashbirdecology` or `@jkarreth`
