### 2021-02-01 #2890 @trashbirdecology
I don’t have the capacity this month to do a timely review.

On Sat, Jan 30, 2021 at 10:05 AM David H Hagan <notifications@github.com>
wrote:

> 👋 @trashbirdecology <https://github.com/trashbirdecology> - it seems
> like you would be a good fit to review this paper. Would be interested and
> have the time to do so?
>
