### 2021-02-15 #2890 @molgor
Hi David,
I’m reviewing  a (JOSS) paper at the moment. I may have time in two weeks, if you haven’t found reviewers by then.

Cheers,

Juan

> El 15/02/2021, a las 08:34, David H Hagan <notifications@github.com> escribió:
>
>
> 👋 Hi there, @rmflight <https://github.com/rmflight> @pboesu <https://github.com/pboesu> @molgor <https://github.com/molgor> @ethanwhite <https://github.com/ethanwhite> @jhollist <https://github.com/jhollist>: do one of you have the time and interest to review this paper?
>
