+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   ambrosia_vignette: ambrosia_vignette.html
-last_built: 2021-03-08T06:06Z
+last_built: 2021-03-10T04:52Z
 
