### 2021-03-10 203ee9a
Documentation updates based on feedback
