+++ man/create_dataset_for_parameter_fit.Rd
@@ -15,6 +15,8 @@ create_dataset_for_parameter_fit(
 )
 }
 \arguments{
+\item{min_price_pd}{Minimum price paid for non-staples.}
+
 \item{min_cal_fd}{Minimum calories for the food demand model}
 
 \item{min_clusters}{Minimum number of clusters to be generated by clustering algoritm. It is recommended to not lowewr this parameter below 20.}
