### 2021-03-10 #2890 @kanishkan91
@dhhagan @jhollist Just an FYI- We did a new release for ambrosia (v1.3.2) with some minor documentation updates and updates to the user tutorial. The above pdf reflects the changes. Please let us know what you think.
