### 2021-03-10 #2890 @jhollist
Perfect timing!  I hope to work on this review today...

On Wed, Mar 10, 2021 at 12:41 AM Kanishka Narayan ***@***.***>
wrote:

> @dhhagan <https://github.com/dhhagan> @jhollist
> <https://github.com/jhollist> Just an FYI- We did a new release for
> ambrosia (v1.3.2) with some minor documentation updates and updates to the
> user tutorial. The above pdf reflects the changes. Please let us know what
> you think.
>
