### 2021-03-15 #2890 @jhollist
@kanishkan91 Just wanted to let you know that I posted my review as an issue in your repository: https://github.com/JGCRI/ambrosia/issues/29.  Details on the unchecked items are in that review plus a few other things.

Overall, nice job!

Any questions let me know.
