### 2021-03-15 12cbc1e
set required R version to >= 3.6
