+++ DESCRIPTION
@@ -7,11 +7,11 @@ Authors@R: c(person(given='Kanishka', family='Narayan', email='kanishka.narayan@
     person(given='James', family='Edmonds', email='jae@pnnl.gov', role=c('aut')),
     person(given='Ryna', family='Cui', email='yiyun.cui@pnnl.gov', role=c('aut')))
 Description: The `ambrosia` R package was developed to calculate food demand for staples and non-staple commodities that is responsive to changing levels of incomes and prices. Ambrosia implements the framework to quantify food demand as established by Edmonds et al. (2017) and allows the user to explore and estimate different variables related to the food demand system.
-Depends: R (>= 3.5.0)
+Depends: R (>= 3.6.0)
 License: BSD2
 Encoding: UTF-8
 LazyData: true
-Imports: 
+Imports:
     dplyr (>= 0.7),
     nleqslv (>= 3.2),
     reshape2 (>= 1.4.3),
@@ -20,7 +20,7 @@ Imports:
     tidyr  (>= 0.7.1),
     rmarkdown
 RoxygenNote: 7.1.1
-Suggests: 
+Suggests:
     shiny (>= 1.0),
     testthat,
     knitr,
