+++ man/create.dataset.for.parameter.fit.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/calculate_parameters.R
-\name{create_dataset_for_parameter_fit}
-\alias{create_dataset_for_parameter_fit}
+\name{create.dataset.for.parameter.fit}
+\alias{create.dataset.for.parameter.fit}
 \title{Create dataset with observational error for log likelihood calculation using clustering.}
 \usage{
-create_dataset_for_parameter_fit(
+create.dataset.for.parameter.fit(
   min_price_pd = 20,
   min_cal_fd = 1000,
   min_clusters = 300,
