+++ man/calculate.ambrosia.params.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/calculate_parameters.R
-\name{calculate_ambrosia_params}
-\alias{calculate_ambrosia_params}
-\title{Calculates the 11 parameters for ambrosia using data calculated by \code{\link{create_dataset_for_parameter_fit()}} by maximizing log likelihood}
+\name{calculate.ambrosia.params}
+\alias{calculate.ambrosia.params}
+\title{Calculates the 11 parameters for ambrosia using data calculated by \code{\link{create.dataset.for.parameter.fit}} by maximizing log likelihood}
 \usage{
-calculate_ambrosia_params(
+calculate.ambrosia.params(
   optim_method = "BFGS",
   original_param_vector = c(1.28, 1.14, -0.19, 0.21, -0.33, 0.5, 0.1, 16, 5.06, 100,
     20),
@@ -20,7 +20,7 @@ calculate_ambrosia_params(
 
 \item{original_param_vector}{Original parameter vector to be used as the starting point for the optimization.These parameters are taken from Edmonds et al 2017}
 
-\item{datadir}{Directory to the data calculated by \code{\link{create_dataset_for_parameter_fit()}}}
+\item{datadir}{Directory to the data calculated by \code{\link{create.dataset.for.parameter.fit}}}
 
 \item{outdir}{Directory to store output csv. Default is set to test_output folder.}
 
@@ -56,7 +56,7 @@ psscl : Price scaling parameter for staples that is applied to the Price of stap
 pnscl : Price scaling parameter for staples that is applied to the Price of non-staples (Pn) and the alpha of non-staples
 }
 \description{
-Calculates the 11 parameters for ambrosia using data calculated by \code{\link{create_dataset_for_parameter_fit()}} by maximizing log likelihood
+Calculates the 11 parameters for ambrosia using data calculated by \code{\link{create.dataset.for.parameter.fit}} by maximizing log likelihood
 }
 \details{
 The following steps are involved in the parameter estimation function.
