+++ paper/paper.bib
@@ -24,7 +24,8 @@
   volume={12},
   number={PNNL-SA-137098},
   year={2019},
-  publisher={Lawrence Berkeley National Lab.(LBNL), Berkeley, CA (United States); Pacific~…}
+  publisher={Lawrence Berkeley National Lab.(LBNL), Berkeley, CA (United States); Pacific~…},
+  DOI={10.5194/gmd-12-677-2019}
 }
 @article{di2020moirai,
   title={Moirai Version 3: A Data Processing System to Generate Recent Historical Land Inputs for Global Modeling Applications at Various Scales},
@@ -91,7 +92,8 @@
   number={1},
   pages={1-9},
   year={2016},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  DOI={10.1038/sdata.2016.18}
 }
 @article{broyden1969,
   title={A new method of solving nonlinear simultaneous equations},
