### 2021-03-17 #2890 @kanishkan91
@jhollist Thank you very much for the detailed review! We have tried to address your review comments as a part of the new release (v1.3.3). The above pdf also reflects the same. Here is a summary of our responses for your convenience.

1. I have added a note in the review stream above about the authorship. As mentioned in the comment above, the other reviewer also had a similar comment. Please let us know what you think.
2. The readme now clarifies that the user can make use of `install_github` directly from `devtools` or from `remotes`. Instructions are added for both. We thought this would be best given that both require `devtools` in any case.
3. We have fixed the links in the documentation and have made sure the Rd errors don't occur anymore. Thank you for the catch.
4. We have added a 'statement of need and target audience' directly in the readme. This is a shortened version from the paper. The version on the readme is more specific as to the need, users. The one in the paper elaborates on the need further. We thought this would be best given that github users can look to the paper if they want more information (e.g. Info on the ecosystem of tools within which ambrosia operates etc.)
5. We have added the DOI's to the references. Thank you very much for the catch!
6.  We ran the CRAN checks using `devtools::check_win_release()`. We have fixed most warnings. The only remaining warnings are about undocumented objects (These are prebuilt datasets that the check is seeing as a code object) and a warning about a GNU extension on a cpp makefile (The documentation clarifies that the makefile is added only for illustrative purposes and is not used within the R code). Let us know what you think of this.
7. Note that we have made sure that the tests also don't give any warnings now.
8. We are now using a consistent left assign in the vignette.
9. We have added a note above the app saying that the user would require `shiny` to run it.
10. We took out the extra comma in the `calculate.ambrosia.params()` function. Thanks for the catch!
11.  The parameter calculation is now faster (though the note clarifies that the run time is different depending on machines). We are trying to control for speed in the example by using a small number of iterations. But the problem is that even with a single iteration `optim()`  (Which is what the estimation function relies on) splits it up into multiples thus increasing run time (It ran on my machine in 15 minutes). Let us know if you have thoughts related to this. Our documentation also mentions that we are using a sample number of iterations in the vignette.
12. We have implemented the changes you suggested to the github actions yml file. We also added the windows release.
13. Function names now use the dot case and not the snake case. Thanks for the catch!
14. We have added the edits to the paper you suggested.
15. We refer to the vignette as `ambroia_vignette` as opposed to `ambrosia_vignette.rmd`
16. The github documentation pages should also be reflective of all of the above.

I know this is a lot, but please do let us know if we are missing anything in the above. Thanks again for this very useful review!
