### 2021-03-17 #2890 @jhollist
@kanishkan91 Thanks for the thorough response!  All this sounds good to me!

Only last suggestion is that I would have you choose `remotes` or `devtools`.  Admittedly, it is a nitpicky thing on my part.  If a user already has installed `devtools` then they will be all set, but if they don't then a `devtools` install has a lot of extra dependencies and `remotes` installs much quicker.  The decision on this one is yours and I'd be happy with either!

Will complete the checklist shortly.

Nice job!
