### 2021-03-17 #2890 @kanishkan91
@jhollist Thanks again. Yes, the `devtools` vs `remotes` is an important one. I guess we can go with `remotes` given that developers will be prompted by R to get `devtools` to add to the package anyway. Other users can just use remotes to quickly install the package. Does this sound ok?

@crvernon Do you have thoughts on this?
