### 2021-03-17 #2890 @jhollist
`devtools` shouldn't be required for any of this from what I can tell.  It isn't a dependency on `remotes` or `ambrosia`.  Would only be around if a user has need for the other development tools (and many users probably would...)

I will leave the decision up to you and your co-authors.  Either option would provide the needed installation tools.
