### 2021-03-17 #2890 @kanishkan91
@jhollist Thanks! Based on the above discussion, we have changed the readme to retain only the `devtools` instructions.

@dhhagan We have addressed review comments from both reviewers. We are thinking we will do one final release 1.3.4 with any other edits that may be required by you. Please let us know what you think of the above and if you need anything else from my side.

Thanks again!
