### 2021-03-22 ecab9e4
Updates for version 1.3.4
