### 2021-03-22 c5376c6
Updates for v1.3.4
