### 2021-03-23 #2890 @crvernon
@dhhagan :

The following is the information you requested:
- GitHub release tag:  https://github.com/JGCRI/ambrosia/releases/tag/v1.3.5
- Zenodo archive reference:
> Kanishka Narayan, Chris R. Vernon, Stephanie Waldhoff, James A. Edmonds, & Ryna Cui. (2021, March 23). ambrosia: An R package for calculating and analyzing food demand that is responsive to changing incomes and prices (Version v1.3.5). Zenodo. http://doi.org/10.5281/zenodo.4628567
- DOI of the archived version:  `10.5281/zenodo.4628567`

Please let us know if you need anything else.

Thanks!
